updates namespace of rules so pattern makes sense in downstream repos
