+++ lib/gl_rubocop/gl_cops/sidekiq_inherits_from_sidekiq_job.rb
@@ -1,0 +1,14 @@
+module GLRubocop
+  module GLCops
+    class SidekiqInheritsFromSidekiqJob < RuboCop::Cop::Cop
+      def on_class(klass)
+        return unless klass.instance_of?(RuboCop::AST::ClassNode)
+        return if klass.parent_class.present? || klass.identifier.short_name == :SidekiqJob
+
+        add_offense(klass)
+      end
+
+      MSG = 'All Sidekiq workers and jobs should inherit from SidekiqJob'.freeze
+    end
+  end
+end
